Corrected Trace
- Now it only expands until collision
- METERS_TO_PIXELs returns ints
- Trace is now an iPoint

+++ Aimbot/EulerIntegrator/Globals.h
@@ -25,7 +25,7 @@ typedef double float64;
 #define PIXELS_PER_METER 50.0f // if touched change METER_PER_PIXEL too
 #define METER_PER_PIXEL 0.02f // this is 1 / PIXELS_PER_METER !
 
-#define METERS_TO_PIXELS(m) ((double) floor(PIXELS_PER_METER * m))
+#define METERS_TO_PIXELS(m) ((int) floor(PIXELS_PER_METER * m))
 #define PIXEL_TO_METERS(p)  ((double) METER_PER_PIXEL * p)
 
 #define DEG_TO_RAD(deg) (deg * 0.0174532925199432957f)
